FRAX-296 Space Bathroom Types Images

+++ src/main/service/space-bathroom-types.service.ts
@@ -7,20 +7,26 @@ import { UserService } from './user.service';
 import { SpaceBathroomTypes } from '../entities/space-bathroom-types.entity';
 import { CreateSpaceBathroomTypesDto } from '../dto/requests/space-bathroom-types/create-space-bathroom-types.dto';
 import { UpdateSpaceBathroomTypesDto } from '../dto/requests/space-bathroom-types/update-space-bathroom-types.dto';
-import { BathroomType } from '../commons/constants/enumerations/space-bathroom-types.enum';
 import { SPACE_BATHROOM_TYPES_RESPONSES } from '../commons/constants/response-constants/space-bathroom-types.constant';
+import { S3UtilsService } from './s3-utils.service';
+import { MEDIA_IMAGE_RESPONSES } from '../commons/constants/response-constants/media-image.constant';
+import { S3 } from 'aws-sdk';
 
 @Injectable()
 export class SpaceBathroomTypesService {
+  private readonly s3 = new S3();
+  private readonly bucketName = process.env.AWS_S3_BUCKET_NAME;
+
   constructor(
     @InjectRepository(SpaceBathroomTypes)
     private readonly spaceBathroomTypesRepository: Repository<SpaceBathroomTypes>,
     private readonly userService: UserService,
+    private readonly s3UtilsService: S3UtilsService,
     private readonly logger: LoggerService,
   ) {}
 
   async findSpaceBathroomTypeByName(
-    name: BathroomType,
+    name: string,
   ): Promise<SpaceBathroomTypes | null> {
     return await this.spaceBathroomTypesRepository.findOne({
       where: { name },
@@ -78,6 +84,7 @@ export class SpaceBathroomTypesService {
 
   async createSpaceBathroomType(
     createSpaceBathroomTypesDto: CreateSpaceBathroomTypesDto,
+    imageFile: Express.Multer.File,
   ): Promise<ApiResponse<SpaceBathroomTypes>> {
     try {
       const existingSpaceBathroomType = await this.findSpaceBathroomTypeByName(
@@ -101,8 +108,25 @@ export class SpaceBathroomTypesService {
       const spaceBathroomType = this.spaceBathroomTypesRepository.create({
         ...createSpaceBathroomTypesDto,
       });
+
       const savedSpaceBathroomType =
         await this.spaceBathroomTypesRepository.save(spaceBathroomType);
+
+      const folderName = 'general_media/images/space_bathroom_types';
+      const fileExtension = imageFile.originalname.split('.').pop();
+      const fileName = `${savedSpaceBathroomType.id}.${fileExtension}`;
+
+      const imageUrlLocation = await this.s3UtilsService.uploadFileToS3(
+        folderName,
+        fileName,
+        imageFile.buffer,
+        imageFile.mimetype,
+      );
+
+      savedSpaceBathroomType.s3_url = imageUrlLocation;
+
+      await this.spaceBathroomTypesRepository.save(savedSpaceBathroomType);
+
       this.logger.log(
         `Space bathroom type ${createSpaceBathroomTypesDto.name} created with ID ${savedSpaceBathroomType.id}`,
       );
@@ -178,6 +202,7 @@ export class SpaceBathroomTypesService {
   async updateSpaceBathroomTypeById(
     id: number,
     updateSpaceBathroomTypesDto: UpdateSpaceBathroomTypesDto,
+    imageFile?: Express.Multer.File,
   ): Promise<ApiResponse<SpaceBathroomTypes>> {
     try {
       const existingSpaceBathroomType =
@@ -198,6 +223,41 @@ export class SpaceBathroomTypesService {
         existingSpaceBathroomType,
         updateSpaceBathroomTypesDto,
       );
+
+      if (imageFile) {
+        const folderName = 'general_media/images/space_bathroom_types';
+        const fileExtension = imageFile.originalname.split('.').pop();
+        const fileName = `${existingSpaceBathroomType.id}.${fileExtension}`;
+        let s3Key = '';
+        let imageUrlLocation = existingSpaceBathroomType.s3_url;
+
+        if (imageUrlLocation) {
+          s3Key = await this.s3UtilsService.extractS3Key(imageUrlLocation);
+        }
+
+        if (s3Key) {
+          if (decodeURIComponent(s3Key) != folderName + '/' + fileName) {
+            const headObject =
+              await this.s3UtilsService.checkIfObjectExistsInS3(s3Key);
+            if (!headObject) {
+              return MEDIA_IMAGE_RESPONSES.MEDIA_IMAGE_NOT_FOUND_IN_AWS_S3(
+                s3Key,
+              );
+            }
+            await this.s3UtilsService.deleteObjectFromS3(s3Key);
+          }
+        }
+
+        imageUrlLocation = await this.s3UtilsService.uploadFileToS3(
+          folderName,
+          fileName,
+          imageFile.buffer,
+          imageFile.mimetype,
+        );
+
+        existingSpaceBathroomType.s3_url = imageUrlLocation;
+      }
+
       await this.spaceBathroomTypesRepository.save(updatedSpaceBathroomType);
       this.logger.log(`Space bathroom type with ID ${id} updated successfully`);
       return SPACE_BATHROOM_TYPES_RESPONSES.SPACE_BATHROOM_TYPE_UPDATED(
@@ -221,10 +281,24 @@ export class SpaceBathroomTypesService {
     try {
       const existingSpaceBathroomType =
         await this.findSpaceBathroomTypeById(id);
+
       if (!existingSpaceBathroomType) {
         return await this.handleSpaceBathroomTypeNotFound(id);
       }
+
+      const s3Key = await this.s3UtilsService.extractS3Key(
+        existingSpaceBathroomType.s3_url,
+      );
+
+      const headObject =
+        await this.s3UtilsService.checkIfObjectExistsInS3(s3Key);
+      if (!headObject) {
+        return MEDIA_IMAGE_RESPONSES.MEDIA_IMAGE_NOT_FOUND_IN_AWS_S3(s3Key);
+      }
+
+      await this.s3UtilsService.deleteObjectFromS3(s3Key);
       await this.spaceBathroomTypesRepository.remove(existingSpaceBathroomType);
+
       this.logger.log(`Space bathroom type with ID ${id} deleted successfully`);
       return SPACE_BATHROOM_TYPES_RESPONSES.SPACE_BATHROOM_TYPE_DELETED(id);
     } catch (error) {
